More warning clean up Follow on to https://gerrit.wikimedia.org/r/#/c/211730/
Down to 18 warnings… so close!

+++ Wikipedia/Categories/UICollectionViewFlowLayout+AttributeUtils.m
@@ -18,8 +18,8 @@
 - (NSArray*)wmf_sortAttributesByLeadingEdgeDistance:(NSArray*)attributes toPoint:(CGPoint)point {
     return [attributes sortedArrayUsingComparator:^NSComparisonResult (UICollectionViewLayoutAttributes* attr1,
                                                                        UICollectionViewLayoutAttributes* attr2) {
-        float leadingEdgeDistance1 = fabsf(point.x - attr1.frame.origin.x);
-        float leadingEdgeDistance2 = fabsf(point.x - attr2.frame.origin.x);
+        float leadingEdgeDistance1 = fabs(point.x - attr1.frame.origin.x);
+        float leadingEdgeDistance2 = fabs(point.x - attr2.frame.origin.x);
         return leadingEdgeDistance1 - leadingEdgeDistance2;
     }];
 }
